requirements.txt
pymupdf
pytesseract
textdistance
streamlit
pandas
jellyfish
pdf2image
openpyxl
pillow
scikit-learn